Extracted functionality out of index.js
