+++ src/actionChecker.js
@@ -18,6 +18,7 @@ export class ActionChecker {
   /**
    * Get abundance service
    * @return {AbundanceService}
+   * @private
    */
   _getAbundanceService () {
     return this.serviceProvider.abundanceService
@@ -26,6 +27,7 @@ export class ActionChecker {
   /**
    * Get fact checker
    * @return {FactChecker}
+   * @private
    */
   _getFactChecker () {
     return this.serviceProvider.factChecker
@@ -34,6 +36,7 @@ export class ActionChecker {
   /**
    * Get context explainer
    * @return {ContextExplainer}
+   * @private
    */
   _getContextExplainer () {
     return this.serviceProvider.contextExplainer
